the fourth section is fully functional

+++ api/delete_hall.php
@@ -6,13 +6,20 @@ if(!isset($_POST['hallID'])) {
     $result = ['status' => false, 'mess' => "Ошибка удаления зала"];
 } else {
     $hallID = (int)($_POST['hallID']);
-    $deleteHall = $db->delete('halls', ['id' => $hallID]);
-
-    if($deleteHall) {
-        $result = ['status' => true];
+    $deleteSeance = $db->delete('seances', ['hall_id' => $hallID]);
+    if($deleteSeance) {
+        $deleteHall = $db->delete('halls', ['id' => $hallID]);
+        if($deleteHall) {
+            $result = ['status' => true];
+        } else {
+            $result = ['status' => false, 'mess' => print_r($deleteHall->errorInfo(), true)];
+        }
     } else {
-        $result = ['status' => false, 'mess' => print_r($deleteHall->errorInfo(), true)];
+        $result = ['status' => false, 'mess' => print_r($deleteSeance->errorInfo(), true)];
     }
+
+
+
 }
 
 print json_encode($result, JSON_UNESCAPED_UNICODE);
